+++ lib/model/weather_pressure.dart
@@ -5,10 +5,10 @@ part 'weather_pressure.g.dart';
 @JsonSerializable()
 class weatherPressure {
   /// 大氣壓力
-  final int pressuremb;
+  final double pressuremb;
 
   /// 大氣壓力
-  final int pressurein;
+  final double pressurein;
 
   const weatherPressure({
     required this.pressuremb,
